--- a/libraries/psibase/native/src/useTriedent.cpp
+++ b/libraries/psibase/native/src/useTriedent.cpp
@@ -308,7 +308,7 @@ namespace psibase
          {
             // std::cout << "Open existing " << dir << "\n";
          }
-         trie   = std::make_shared<triedent::database>(dir.c_str(), triedent::database::read_write);
+         trie = std::make_shared<triedent::database>(dir.c_str(), triedent::open_mode::read_write);
          auto s = trie->start_write_session();
          head   = loadRevision(*s, s->get_top_root(), revisionHeadKey);
       }