|| before | after |
| - | - | - |
| t-Nft-test | 33.11 sec | 1.65 sec |
|t-Symbol-test | 34.89 sec | 2.08 sec |
|t-Tokens-test| 90.30 sec | 2.53 sec |
|t-psibase-tests | 9.84 sec | 1.63 sec |